Add MIT license
Adding the MIT License to clarify the terms under which this project can be used, modified, and distributed. This open-source license allows others to freely use, copy, modify, and distribute the code while providing a disclaimer of liability. This commit is intended to make the project publicly accessible with clear usage guidelines.